Refactor initial files
