+++ pyproject.toml
@@ -17,6 +17,8 @@ jsonschema = "4.4.0"
 jsonpath-rw = "^1.4.0"
 jsonpath-rw-ext = "^1.2.2"
 eq-translations = {git = "https://github.com/ONSDigital/eq-translations.git", rev = "v4.9.2"}
+fastapi = "^0.113.0"
+uvicorn = "^0.30.6"
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.4.2"
